requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
pillow
pyyaml
dvclive
dvc[ssh]
matplotlib
tqdm
